move fixedpoint types to their own file
