+++ lib/ethpy/ethpy/hyperdrive/state/__init__.py
@@ -1,6 +1,5 @@
 """Hyperdrive state classes and conversion helper functions."""
+from hypertypes.fixedpoint_types import FeesFP, PoolConfigFP, PoolInfoFP
+
 from .checkpoint import Checkpoint
-from .fees import Fees
-from .pool_config import PoolConfig
-from .pool_info import PoolInfo
 from .pool_state import PoolState
