+++ lib/ethpy/ethpy/hyperdrive/transactions.py
@@ -5,6 +5,7 @@ from eth_typing import BlockNumber
 from ethpy.base import UnknownBlockError, get_transaction_logs
 from fixedpointmath import FixedPoint
 from hypertypes import IERC4626HyperdriveContract
+from hypertypes.fixedpoint_types import PoolConfigFP, PoolInfoFP
 from hypertypes.utilities.conversions import (
     hypertypes_checkpoint_to_fixedpoint,
     hypertypes_pool_config_to_fixedpoint,
@@ -14,10 +15,10 @@ from web3.contract.contract import Contract
 from web3.types import Timestamp, TxReceipt
 
 from .receipt_breakdown import ReceiptBreakdown
-from .state import Checkpoint, PoolConfig, PoolInfo
+from .state import Checkpoint
 
 
-def get_hyperdrive_pool_config(hyperdrive_contract: IERC4626HyperdriveContract) -> PoolConfig:
+def get_hyperdrive_pool_config(hyperdrive_contract: IERC4626HyperdriveContract) -> PoolConfigFP:
     """Get the hyperdrive config from a deployed hyperdrive contract.
 
     Arguments
@@ -34,7 +35,7 @@ def get_hyperdrive_pool_config(hyperdrive_contract: IERC4626HyperdriveContract) 
     return hypertypes_pool_config_to_fixedpoint(pool_config)
 
 
-def get_hyperdrive_pool_info(hyperdrive_contract: IERC4626HyperdriveContract, block_number: BlockNumber) -> PoolInfo:
+def get_hyperdrive_pool_info(hyperdrive_contract: IERC4626HyperdriveContract, block_number: BlockNumber) -> PoolInfoFP:
     """Get the block pool info from the Hyperdrive contract.
 
     Arguments
