+++ lib/hypertypes/hypertypes/__init__.py
@@ -1,2 +1,3 @@
 """Top level hypertypes init file."""
+from .fixedpoint_types import *
 from .types import *
