+++ lib/hypertypes/hypertypes/fixedpoint_types.py
@@ -1,9 +1,11 @@
-"""Functions for storing Hyperdrive state."""
+"""Fixed point versions of common structs from the hyperdrive contracts."""
 from __future__ import annotations
 
 from dataclasses import dataclass
+from typing import Sequence
 
 from fixedpointmath import FixedPoint
+
 
 # TODO: These dataclasses are similar to pypechain except for
 #  - snake_case attributes instead of camelCase
@@ -15,10 +17,17 @@ from fixedpointmath import FixedPoint
 # One solution could be to build our own interface wrapper that pulls in the pypechain
 # dataclass and makes this fixed set of changes?
 # pylint: disable=too-many-instance-attributes
+@dataclass
+class FeesFP:
+    """Fees struct."""
+
+    curve: FixedPoint
+    flat: FixedPoint
+    governance: FixedPoint
 
 
 @dataclass
-class PoolInfo:
+class PoolInfoFP:
     """PoolInfo struct."""
 
     share_reserves: FixedPoint
@@ -34,3 +43,28 @@ class PoolInfo:
     withdrawal_shares_proceeds: FixedPoint
     lp_share_price: FixedPoint
     long_exposure: FixedPoint
+
+
+@dataclass
+class PoolConfigFP:
+    """PoolConfig struct."""
+
+    base_token: str
+    linker_factory: str
+    linker_code_hash: bytes
+    initial_share_price: FixedPoint
+    minimum_share_reserves: FixedPoint
+    minimum_transaction_amount: FixedPoint
+    precision_threshold: int
+    position_duration: int
+    checkpoint_duration: int
+    time_stretch: FixedPoint
+    governance: str
+    fee_collector: str
+    # TODO: Pyright:
+    # Declaration "fees" is obscured by a declaration of the same name here but not elsewhere
+    fees: FeesFP | Sequence  # type: ignore
+
+    def __post_init__(self):
+        if isinstance(self.fees, Sequence):
+            self.fees: FeesFP = FeesFP(*self.fees)
